deploy: updating Windows choco service path logic for Program Files
Summary:
Due to a [privilege escalation in osquery](https://www.facebook.com/security/advisories/cve-2019-3567) being located in
`ProgramData`, we require installs "pivot" to `Program Files` to prevent
this vulnerability. In doing so we did not account for updating any
pre-existing service paths during this install. This logic ensures that
if the system service already exists, and points to the old location, we
update the reference to the new location.

We do not remove the legacy path installation as it may be the case that
users will need to migrate databases or other system configurations, as
such it is left to the deployment to clean up the `ProgramData`
installation as it will no longer be referenced.

Test Plan:
Will update with sample installations both clean and with an in place
install running.